requirements.txt
Django>=5.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
gunicorn>=23.0.0
django-ninja>=1.4.0
django-cors-headers>=4.3.0
whitenoise>=6.9.0
pandas>=2.3.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.30.0
aiohttp>=3.9.0

